# perf: clean up proto and improve mesgdef