--- a/internal/cmd/fitgen/profile/mesgdef/mesgdef.tmpl
+++ b/internal/cmd/fitgen/profile/mesgdef/mesgdef.tmpl
@@ -66,33 +66,28 @@ func New{{ .Name }}(mesg *proto.Message) *{{ .Name }} {
 
 // ToMesg converts {{ .Name }} into proto.Message.
 func (m *{{ .Name }}) ToMesg(fac Factory) proto.Message {
+    fieldsPtr := fieldsPool.Get().(*[256]proto.Field)
+    defer fieldsPool.Put(fieldsPtr)
+    
+    fields := (*fieldsPtr)[:0] // Create slice from array with zero len.
     mesg := fac.CreateMesgOnly(typedef.MesgNum{{ .Name }})
-	mesg.Fields = make([]proto.Field, 0, m.size())
-
+	
     {{ range .Fields -}}
         if {{ .ComparableValue }} != {{ .InvalidValue }} {
             field := fac.CreateField(mesg.Num, {{ .Num }})
             field.Value = {{ .PrimitiveValue }}
-            mesg.Fields = append(mesg.Fields, field)
+            fields = append(fields, field)
         }
 	{{ end }}
+
+    mesg.Fields = make([]proto.Field, len(fields))
+    copy(mesg.Fields, fields)
 
 	{{ if and (not (eq .Name "FileId")) (not (eq .Name "DeveloperDataId")) (not (eq .Name "FieldDescription"))  }}
 	mesg.DeveloperFields = m.DeveloperFields
 	{{ end }}
 
 	return mesg
-}
-
-// size returns size of {{ .Name }}'s valid fields.
-func (m *{{ .Name }}) size() byte {
-    var size byte
-    {{ range .Fields -}}
-        if {{ .ComparableValue }} != {{ .InvalidValue }} {
-            size++
-        }
-	{{ end -}}
-    return size
 }
 
 {{ range .Fields -}}