--- a/profile/mesgdef/accelerometer_data_gen.go
+++ b/profile/mesgdef/accelerometer_data_gen.go
@@ -72,115 +72,79 @@ func NewAccelerometerData(mesg *proto.Message) *AccelerometerData {
 
 // ToMesg converts AccelerometerData into proto.Message.
 func (m *AccelerometerData) ToMesg(fac Factory) proto.Message {
+	fieldsPtr := fieldsPool.Get().(*[256]proto.Field)
+	defer fieldsPool.Put(fieldsPtr)
+
+	fields := (*fieldsPtr)[:0] // Create slice from array with zero len.
 	mesg := fac.CreateMesgOnly(typedef.MesgNumAccelerometerData)
-	mesg.Fields = make([]proto.Field, 0, m.size())
 
 	if datetime.ToUint32(m.Timestamp) != basetype.Uint32Invalid {
 		field := fac.CreateField(mesg.Num, 253)
 		field.Value = datetime.ToUint32(m.Timestamp)
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.TimestampMs != basetype.Uint16Invalid {
 		field := fac.CreateField(mesg.Num, 0)
 		field.Value = m.TimestampMs
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.SampleTimeOffset != nil {
 		field := fac.CreateField(mesg.Num, 1)
 		field.Value = m.SampleTimeOffset
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.AccelX != nil {
 		field := fac.CreateField(mesg.Num, 2)
 		field.Value = m.AccelX
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.AccelY != nil {
 		field := fac.CreateField(mesg.Num, 3)
 		field.Value = m.AccelY
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.AccelZ != nil {
 		field := fac.CreateField(mesg.Num, 4)
 		field.Value = m.AccelZ
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.CalibratedAccelX != nil {
 		field := fac.CreateField(mesg.Num, 5)
 		field.Value = m.CalibratedAccelX
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.CalibratedAccelY != nil {
 		field := fac.CreateField(mesg.Num, 6)
 		field.Value = m.CalibratedAccelY
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.CalibratedAccelZ != nil {
 		field := fac.CreateField(mesg.Num, 7)
 		field.Value = m.CalibratedAccelZ
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.CompressedCalibratedAccelX != nil {
 		field := fac.CreateField(mesg.Num, 8)
 		field.Value = m.CompressedCalibratedAccelX
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.CompressedCalibratedAccelY != nil {
 		field := fac.CreateField(mesg.Num, 9)
 		field.Value = m.CompressedCalibratedAccelY
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
 	if m.CompressedCalibratedAccelZ != nil {
 		field := fac.CreateField(mesg.Num, 10)
 		field.Value = m.CompressedCalibratedAccelZ
-		mesg.Fields = append(mesg.Fields, field)
+		fields = append(fields, field)
 	}
+
+	mesg.Fields = make([]proto.Field, len(fields))
+	copy(mesg.Fields, fields)
 
 	mesg.DeveloperFields = m.DeveloperFields
 
 	return mesg
-}
-
-// size returns size of AccelerometerData's valid fields.
-func (m *AccelerometerData) size() byte {
-	var size byte
-	if datetime.ToUint32(m.Timestamp) != basetype.Uint32Invalid {
-		size++
-	}
-	if m.TimestampMs != basetype.Uint16Invalid {
-		size++
-	}
-	if m.SampleTimeOffset != nil {
-		size++
-	}
-	if m.AccelX != nil {
-		size++
-	}
-	if m.AccelY != nil {
-		size++
-	}
-	if m.AccelZ != nil {
-		size++
-	}
-	if m.CalibratedAccelX != nil {
-		size++
-	}
-	if m.CalibratedAccelY != nil {
-		size++
-	}
-	if m.CalibratedAccelZ != nil {
-		size++
-	}
-	if m.CompressedCalibratedAccelX != nil {
-		size++
-	}
-	if m.CompressedCalibratedAccelY != nil {
-		size++
-	}
-	if m.CompressedCalibratedAccelZ != nil {
-		size++
-	}
-	return size
 }
 
 // SetTimestamp sets AccelerometerData value.