… profile/mesgdef/bike_profile_gen.go
… profile/mesgdef/camera_event_gen.go
… profile/mesgdef/capabilities_gen.go
… profile/mesgdef/climb_pro_gen.go
… profile/mesgdef/connectivity_gen.go
… profile/mesgdef/course_gen.go
… profile/mesgdef/course_point_gen.go
… profile/mesgdef/developer_data_id_gen.go
… profile/mesgdef/device_aux_battery_info_gen.go
… profile/mesgdef/device_info_gen.go
… profile/mesgdef/device_settings_gen.go
… profile/mesgdef/dive_alarm_gen.go
… profile/mesgdef/dive_apnea_alarm_gen.go
… profile/mesgdef/dive_gas_gen.go
… profile/mesgdef/dive_settings_gen.go
… profile/mesgdef/dive_summary_gen.go
… profile/mesgdef/event_gen.go
… profile/mesgdef/exd_data_concept_configuration_gen.go
… profile/mesgdef/exd_data_field_configuration_gen.go
… profile/mesgdef/exd_screen_configuration_gen.go
… profile/mesgdef/exercise_title_gen.go
… profile/mesgdef/field_capabilities_gen.go
… profile/mesgdef/field_description_gen.go
… profile/mesgdef/file_capabilities_gen.go
… profile/mesgdef/file_creator_gen.go
… profile/mesgdef/file_id_gen.go
… profile/mesgdef/goal_gen.go
… profile/mesgdef/gps_metadata_gen.go
… profile/mesgdef/gyroscope_data_gen.go
… profile/mesgdef/hr_gen.go