- clean up proto.Field's Size since no longer used
- improve mesgdef, now use fields pool, a backed array for slicing, instead of counting size before creating the slice.

benchstat:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/profile/mesgdef
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │ record_gen_old.txt │       record_gen_new.txt       │
         │       sec/op       │    sec/op     vs base          │
ToMesg-4         968.3n ± 24%   959.3n ± 20%  ~ (p=0.912 n=10)

         │ record_gen_old.txt │        record_gen_new.txt         │
         │        B/op        │    B/op     vs base               │
ToMesg-4           360.0 ± 0%   352.0 ± 0%  -2.22% (p=0.000 n=10)

         │ record_gen_old.txt │         record_gen_new.txt         │
         │     allocs/op      │ allocs/op   vs base                │
ToMesg-4          11.000 ± 0%   9.000 ± 0%  -18.18% (p=0.000 n=10)

```